feat(generator): export from json to html - add: tools/htmlToHeader to create raw string from html/js/css triplet in build time - chore: integrate it in cmake build pipeline - add: src/generator/exporter.hpp - chore: rename generator into jsonGenerator - add: basic html/css/js template
test(generator): test export and astToJson logic
- add: test_jsonGenerator
- add: test_exporter

fix: cedictdatasize type to ulonglong

fix(jsonGenerator): id count not cleared if astToJson called multiple time
- add: id in Module class and remove it from json generation
- chore: adapt parser.cpp

perf(tools): improve binToHeader tool by avoiding io operations in for loop
and added errors messages for security.

+++ src/enricher/cedict.cpp
@@ -4,7 +4,7 @@
 
 extern const unsigned char cedictData[];
 extern const unsigned int cedictEntriesCount;
-extern const unsigned int cedictDataSize;
+extern const unsigned long long cedictDataSize;
 
 struct KeyCompHanzi
 {
